+++ ForSubmit/Makefile
@@ -6,7 +6,7 @@
 #    By: yokitaga <yokitaga@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/10/08 11:19:11 by yokitaga          #+#    #+#              #
-#    Updated: 2023/11/26 22:54:25 by yokitaga         ###   ########.fr        #
+#    Updated: 2023/11/28 15:13:32 by yokitaga         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -20,6 +20,7 @@ SRCS 			= $(addprefix $(SRCS_DIR),\
 			0_main.c		\
 			1_check_arg.c		\
 			2_read_map.c		\
+			3_0_parse_read_data.c \
 			3_1_parse_read_data.c \
 			3_2_parse_read_data.c \
 			3_3_parse_read_data.c \
